requirements.txt
beautifulsoup4~=4.12.3
selenium~=4.23.1
colorlog~=6.8.2
Unidecode~=1.3.8
pandas~=2.2.3
dateparser~=1.2.0
requests~=2.32.3
python-dotenv~=1.0.1